feat: lazy-load charts, avoid ssr
